requirements.txt
appdirs==1.4.3
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
djangorestframework-jsonp==1.0.2
djangorestframework-recursive==0.1.1
packaging==16.8
pygame==1.9.3
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
